Create example personas and add infrastructure (#2)
This PR adds the basic infrastructure and one or two example personas.
More personas will be added later.

The infrastructure follows this logic:
- Personas can be created for a specific date. This is to account for
different inputs for different policy regimes in GETTSIM.
- Users call `personas_for_date` and get a `PersonaCollection` object
out of it. Personas are attributes of this object. When doing this I
head the optimagic experience in mind when users are selection
algorithms (the IDE lists the available options in a context menu). This
is exactly how it works now with personas.

---------

Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>