Upserting input data based on personas (#5)
### What problem do you want to solve?

This PR provides a function to let the user upsert data into
NestedData, typically provided by some Persona.

Data provided by the user needs to have the length of a multiple of the
original (base) data. All data columns that are not upserted by the user
are broadcasted to match the length of upserted data, including foreign
keys and group variables.